+++ scaleph-plugins/scaleph-plugin-seatunnel-native-flink/src/main/java/cn/sliew/scaleph/plugin/seatunnel/flink/jdbc/source/plugin.properties
@@ -1,0 +1,4 @@
+name=SeatunnelNativeFlinkJdbcSource
+description=seatunnel jdbc source based on native flink
+version=1.0.0
+classname=cn.sliew.scaleph.plugin.seatunnel.flink.jdbc.source.JdbcSource
